Decoder_ALLOC_SAMPLE {BRKGA}	R Documentation

Decoder implemented to optimal allocation in stratified sampling

Description

Builds a solution associated with sample sizes allocated by stratum

Usage


Decoder_ALLOC_SAMPLE(u, Nh)

Arguments

u
chromossome with H position (number of strata)

Nh
total Units in hth stratum (is necessary to define Exa1 = Nh)

Details

Decoder_ALLOC_SAMPLE

Value

nh
= number of units in the sample in the hth stratum

Examples


H<-3
u<-popgen(n=H,p=1)
Nh<-c(212,84,61)
nh<-Decoder_ALLOC_SAMPLE(u,Nh)

[Package BRKGA version 0.1.0 Index]