Decoder_MDP {BRKGA}	R Documentation

Decoder implemented to maximum diversity problem

Description

Builds a feasible solution associated with a subset M of N

Usage


Decoder_MDP(u, OD)

Arguments

u
Chromossome

OD
List with Ratio distance matrix and cardinality of M (is necessary to define Exa1 = OD)

Details

Decoder_MDP

Value

Xm
= elements of M

Examples


data(Data_MDP1)
D<-Data_MDP1
N<-dim(D)[1]
M<-10
RDi<-apply(as.matrix(1:N),1,function(i) sum(D[i,]))/sum((D[upper.tri(D)]))
u<-popgen(n=N,1)
OD<-list(M=M,RDi=RDi)
Xm<-Decoder_MDP(u,OD)

[Package BRKGA version 0.1.0 Index]