Fobj_ALLOC_SAMPLE {BRKGA}	R Documentation

Objective function of the optimal allocation in stratified sampling

Description

Calculates sum of sample sizes associated with all strata

Usage


Fobj_ALLOC_SAMPLE(Sh2, nh, S)

Arguments

Sh2
Variance by stratum

nh
Number of units in the sample in the hth stratum

S
List with Nh, Y and cvt(cv target) (is necessary to define Exa2=S)

Details

Fobj_ALLOC_SAMPLE

Value

n
= sum of nh

References

1. Brito, J. A. M.; Silva, P. L. N. ; Semaan, G. S. ; Maculan, N.(2015). Integer Programming Formulations Applied to Optimal Allocation in Stratified Sampling. Survey Methodology, v. 41, p. 427-442.

Examples


Nh<-c(212,84,61)
Sh2<-c(723.1,2693.4,36231.7)
Y<-80548
H<-3
S<-list(Nh=Nh,Y=Y,cvt=0.05)
nh<-c(150,42,30)
n<-Fobj_ALLOC_SAMPLE(Sh2,nh,S)

[Package BRKGA version 0.1.0 Index]