Fobj_MDP {BRKGA}	R Documentation

Objective function of the maximum diversity problem

Description

Calculates the sum of all distances associated with M

Usage


Fobj_MDP(D, Xm)

Arguments

D
distance matrix

Xm
elements of M

Details

Fobj_MDP

Value

SD
= sum all distances associated with M

References

1. Martí, R., Gallego, M., Duarte, A. e Pardo, E. G. (2013) Heuristics and metaheuristics for the maximum diversity problem. Journal of Heuristics 19(4): 591-615.

2. Silva, G.C., Ochi, L.S. e Martins, S.L. (2004). Experimental Comparison of Greedy Randomized Adaptive Search Procedures for the Maximum Diversity Problem. In Experimental and Efficient Algorithms, volume 3059 of Lecture Notes in Computer Science, p. 498b-512. Springer Berlin / Heidelberg.

Examples


data(DataMDP1)
D<-Data_MDP1
N<-dim(D)[1]
M<-10
xm<-sample(N,M)
SD<-Fobj_MDP(D,xm)

[Package BRKGA version 0.1.0 Index]