brkga {BRKGA}	R Documentation

BRKGA Algorithm

Description

This function applies brkga algorithm to solve a problem considering objective function and decoder defined by user

Usage


brkga(
  Data,
  Fo,
  Dc,
  rc = 0.7,
  pe = 0.2,
  pm = 0.2,
  n,
  p = 100,
  ng = 2000,
  ngw = 500,
  MaxTime = 3600,
  MAX = FALSE,
  Exa1 = NULL,
  Exa2 = NULL
)

Arguments

Data
Vector or Matrix (data to calculate objective function)

Fo
Objective function defined by user

Dc
Decoder defined by user

rc
Crossover probability

pe
Percentual of elite chromosomes

pm
Percentual of mutant chromosomes

n
Number of genes in the chromosomes associated with a solution

p
Number of elements (chromosomes) in the population

ng
Number of generations of the brkga algorithm

ngw
Number of generations without improvement

MaxTime
Maximum CPU Time (seconds)

MAX
Argument that determines maximization (TRUE) or minimization problem (FALSE)

Exa1
Extra Argument to Decoder

Exa2
Extra Argument to Objective Function

Details

brkga

Value

fbest
Best value of Objective Function

gbest
Best Solution

cpu_time
Cpu time in seconds

ibest
Generation in which the best solution was found

Author(s)

Jose Brito (jambrito@gmail.com), Gustavo Semaan (gustavosemaan@gmail.com) and Augusto Fadel (augustofadel@gmail.com).

References

1. Gonçalves, J.F. and Resende, M.G.C. (2011). Biased random-key genetic algorithms for combinatorial optimization, Journal of Heuristics, 17, p 487-525.

2. Gonçalves, J.F and Resende M.G.C. (2018). Biased random-key genetic programming. Handbook of Heuristics, Edited by: MartC-, R., Pardalos, P.M. Pardalos and Resende M.G.C. , pp. 23-37, Springer.

Examples


#Example1 - TSP - Travelling Salesman Problem
data(Data_capitals)
D<-as.matrix(dist(Data_capitals))
Ncities<-nrow(Data_capitals)
s<-brkga(Data=D,Fo=Fobj_TSP,Dc=Decoder_TSP,ng=20,n=Ncities,p=50)

#Example2 - Clustering Problem: k-medois
Distance<-as.matrix(dist(iris[,1:4]))
N<-nrow(iris)
k<-3 #Clusters
s<-brkga(Data=Distance,Fo=Fobj_MEDOID,Dc=Decoder_MEDOID,n=N,p=25,ng=20,MaxTime=2,Exa1=k,Exa2=k)

#Example3 - Knapsack Problem
wi<-c(40,50,30,10,10,40,30)
li<-c(40,80,10,10,4,20,60)
C<-100
Datalw<-cbind(li,wi)
s<-brkga(Data=Datalw,Fo=Fobj_KNAPSACK,Dc=Decoder_KNAPSACK,n=length(wi),p=50,ng=20,Exa2=C,MAX=TRUE)

wi<-c(8,10,39,94,32,88,64,90,20,63,71,7,17,99,77,57,3,26,43,55,95,53,34,62,74)
li<-c(23,18,40,87,34,27,44,46,47,65,90,26,6,22,31,12,10,54,71,36,30,63,59,45,79)
C<-400
Datalw<-cbind(li,wi)
s<-brkga(Data=Datalw,Fo=Fobj_KNAPSACK,Dc=Decoder_KNAPSACK,n=length(wi),ng=50,ngw=25,Exa2=C,MAX=TRUE)

#Example4 - MDP Problem - Maximum Diversity Problem
data(Data_MDP1)
D<-Data_MDP1
N<-dim(D)[1]
M<-10
RDi<-apply(as.matrix(1:N),1,function(i) sum(D[i,]))/sum((D[upper.tri(D)]))
Exa1<-list(M=M,RDi=RDi)
s<-brkga(Data=D,Fo=Fobj_MDP,Dc=Decoder_MDP,n=N,ng=10,p=50,MaxTime=2,Exa1=Exa1,MAX=TRUE)

#Example5 - Optimal Allocation  in Stratified Sampling
Nh<-c(212,84,61)
Sh2<-c(723.1,2693.4,36231.7)
Y<-80548
H<-3
Exa2<-list(Nh=Nh,Y=Y,cvt=0.05)
s<-brkga(Data=Sh2,Fo=Fobj_ALLOC_SAMPLE,Dc=Decoder_ALLOC_SAMPLE,n=H,p=1000,ng=50,
MaxTime=2,Exa1=Nh,Exa2=Exa2,MAX=FALSE)

#Example6 - Minimization of Himmelblaus Function
nv<-2 #Number of variables
s<-brkga(Data=NULL,Fo=Fobj_HIMMELBLAUS,Dc=Decoder_HIMMELBLAUS,n=nv,p=100,ng=200)

#Example7 - MSDCP - Minimum Sum Distance Clustering Problem
k<-3 #Clusters
Distance<-as.matrix(dist(iris[,1:4]))
N<-nrow(iris)
s<-brkga(Data=Distance,Fo=Fobj_MSDCP,Dc=Decoder_MSDCP,n=N,p=50,ng=10,MaxTime=2,Exa1=k,Exa2=k)

[Package BRKGA version 0.1.0 Index]