crossover {BRKGA}	R Documentation

Crossover Operator

Description

This function performs the uniform crossover between two chromosomes

Usage


crossover(gelite, gnelite, rc, p, pe, pm, n)

Arguments

gelite
matrix with elite chromosomes

gnelite
matrix with nonelite chromosomes

rc
crossover probability

p
number of elements (chromosomes) in the population

pe
number of elite chromosomes

pm
number of mutant chromossomes

n
number of genes in the chromosomes associated with a solution

Details

crossover

Value

gnew
= matrix pxn with chromosomes produced by crossover

References

1. Gonçalves, J.R. and Resende, M.G.C. (2011). Biased random-key genetic algorithms for combinatorial optimization,

2. Spears, W.M., DeJong, K.A.(1991). On the virtues of parameterized uniform crossover. In: Proceedings of the Fourth International Conference on Genetic Algorithms, pp. 230b.

Examples


X<-popgen(10,10)
Y<-crossover(gelite=X[1:3,],gnelite=X[4:10,],rc=0.7,p=10,pe=3,pm=3,n=10)

[Package BRKGA version 0.1.0 Index]